+++ HISTORY.rst
@@ -8,12 +8,13 @@ History
 * !92: Fixed #111 (DeprecationWarning: `interp2d` is deprecated!).
 * !93: Added a warning if the input dataset was processed with an old version of the ground segment processing system
   (closes #129).
+* !94: Improved all tests for the CLI and the Orthorectifier module that are related to the target projection type.
 
 
 0.19.5 (2023-12-11)
 -------------------
 
-* Added missing EnMAP-Box requirement h5py to  environment files.
+* Added missing EnMAP-Box requirement h5py to environment files.
 * Replaced deprecated direct calls of setup.py within Makefile.
 * Replaced deprecated URL.
 * !75: Disabled dead pixel correction by default and changed the default interpolation approach from spectral to spatial